requirements.txt
# pip install -r requirements.txt

flask
google.generativeai
langchain-groq
streamlit